requirements.txt
streamlit
opperai
pydantic
speechrecognition
pyttsx3
pandas
openpyxl
twilio
python-dotenv
requests
